Move binding ServerIdleDetctor to the assembly module
