+++ wsmaster/che-core-api-workspace/pom.xml
@@ -218,6 +218,18 @@
     </dependencies>
     <build>
         <plugins>
+            <plugin>
+                <groupId>com.mycila</groupId>
+                <artifactId>license-maven-plugin</artifactId>
+                <configuration>
+                    <excludes>
+                        <!-- Exclude files until #3281 is resolved -->
+                        <exclude>**/ServerIdleDetector.java</exclude>
+                        <exclude>**/ServerIdleEvent.java</exclude>
+                        <!-- End excluded files -->
+                    </excludes>
+                </configuration>
+            </plugin>
             <!-- Create the test jar -->
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
